Made the intro page only load on desktops
Set the intro page to skip if the screen resolution is detected below
1200 px.
Also made the intro photos turn grey on hover so that it's a bit more
obvious for users that they need to click the image. Couldn't quite get
it to turn grey on the image map, so had to settle for the entire
rectangular image.